requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
docopt==0.6.2
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
openai==0.27.0
pipreqs==0.4.11
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.4
yarg==0.1.9
yarl==1.9.2
